// src/models/IoK8sApiAdmissionregistrationV1alpha1ParamRef/types.ts
import { IoK8sApimachineryPkgApisMetaV1LabelSelectorModelOptions as IoK8sApimachineryPkgApisMetaV1LabelSelector } from '../IoK8sApimachineryPkgApisMetaV1LabelSelector';

export interface IoK8sApiAdmissionregistrationV1alpha1ParamRefModelOptions {
  name?: string;
  namespace?: string;
  parameterNotFoundAction?: string;
  selector?: IoK8sApimachineryPkgApisMetaV1LabelSelector;
}